refactor: update ArduinoDomain import path and add Bluetooth dependencies
This commit updates the import path for `ArduinoDomain` model across multiple modules.
It also introduces Bluetooth permissions in `AndroidManifest.xml` and adds the `android-bluetooth` dependency in `gradle/libs.versions.toml`.

Additionally, the `settings.gradle.kts` file has been updated to include new modules:
- `:data:connectionBluetooth`
- `:domain:connectionBluetooth`
- `:domain:core`
