+++ data/conectionInternet/src/commonMain/kotlin/com/virtualworld/conectioninternet/data/conectionInternet/RemoteDataSource.kt
@@ -4,7 +4,7 @@ import com.virtualworld.multiplatformiot.ProductEmptyException
 import com.virtualworld.multiplatformiot.data.core.NetworkResponseState
 import com.virtualworld.multiplatformiot.data.core.dto.ArduinoData
 import com.virtualworld.multiplatformiot.data.core.dto.StateObject
-import com.virtualworld.multiplatformiot.domain.conectionInternet.model.ArduinoDomain
+import com.virtualworld.multiplatformiot.domain.core.models.ArduinoDomain
 import dev.gitlive.firebase.firestore.FirebaseFirestore
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
