+++ domain/conectionInternet/src/commonMain/kotlin/com/virtualworld/multiplatformiot/domain/conectionInternet/repository/RepositoryInternet.kt
@@ -3,7 +3,7 @@ package com.virtualworld.multiplatformiot.domain.conectionInternet.repository
 import com.virtualworld.multiplatformiot.data.core.NetworkResponseState
 import com.virtualworld.multiplatformiot.data.core.dto.ArduinoData
 import com.virtualworld.multiplatformiot.data.core.dto.StateObject
-import com.virtualworld.multiplatformiot.domain.conectionInternet.model.ArduinoDomain
+import com.virtualworld.multiplatformiot.domain.core.models.ArduinoDomain
 import kotlinx.coroutines.flow.Flow
 
 interface RepositoryInternet {
